+++ src/app/styles/form-user-input-style/form-user-input-style.component.ts
@@ -85,7 +85,7 @@ export class FormUserInputStyleComponent extends BasicStyleComponent implements 
     public async submitForm(value: { [key: string]: any; }) {
         const res = await this.selfhelpService.submitForm(this.url, this.prepareParams(value));
         if (res && this.getFieldContent('close_modal_at_end') == '1') {
-            this.selfhelpService.closeModal();
+            this.selfhelpService.closeModal('submit');
         }
         if (this.getFieldContent('redirect_at_end') != '') {
             this.selfhelpService.openUrl(this.getFieldContent('redirect_at_end'));
@@ -97,7 +97,7 @@ export class FormUserInputStyleComponent extends BasicStyleComponent implements 
         params['btnSubmitAndSend'] = 'send_email';
         const res = await this.selfhelpService.submitForm(this.url, params);
         if (res && this.getFieldContent('close_modal_at_end') == '1') {
-            this.selfhelpService.closeModal();
+            this.selfhelpService.closeModal('submit');
         }
         if (this.getFieldContent('redirect_at_end') != '') {
             this.selfhelpService.openUrl(this.getFieldContent('redirect_at_end'));