+++ src/app/styles/survey-js-style/survey-js-style.component.ts
@@ -101,7 +101,7 @@ export class SurveyJSStyleComponent extends BasicStyleComponent implements OnIni
 
     private endSurvey() {
         if (this.getFieldContent('close_modal_at_end') == '1') {
-            this.selfhelpService.closeModal();
+            this.selfhelpService.closeModal('submit');
             if (this.getFieldContent('redirect_at_end') != '') {
                 this.selfhelpService.openUrl(this.getFieldContent('redirect_at_end'));
             } else {